server/requirements.txt
python-dotenv>=1.0.0
psutil>=5.9.0
supabase>=2.0.0
flask>=3.0.0
